Feat: ADD Activity, Reward, Quest
- 퀘스트, 보상, 활동기록 추가
- 초대 시 알림 이벤트 발생 -> 알림 생성 + 활동기록 + 보상Trigger

- 배지 추가 예정

+++ src/main/java/kr/co/jeelee/kiwee/domain/member/entity/Member.java
@@ -134,7 +134,7 @@ public class Member extends BaseTimeEntity {
 		this.avatarUrl = avatarUrl;
 	}
 
-	public void gainExp(long exp) {
+	public void gainExp(int exp) {
 		if (exp < 0) {
 			throw new FieldValidationException("exp", "받는 경험치는 음수가 될 수 없습니다.");
 		}
